refactor: extract account mapping model hook
